Add support for multi-validators (#17)
* Add support for multi-validators

The idea is to make it possible to combine multiple validators together
but still support `.on` and `.keys` on the combined validator. This
makes it possible to arbitrarily compose validators transparently.

This commit also firms up the design of custom validation DSLs, so you
could make your own classes like the multi() validator. However, most of
the time, you should be able to use regular functions for composition.
The only reason multi() was needed was to proxy through .on and .keys to
the individual validators.

* 0.5.6

* Make the multi() DSL non-mutating

* 0.5.7