[BREAKING] Refactor away from Label -> Descriptor
All remaining usage of Label can be migrated to the new descriptor
representation. Also migrated alias concepts to an explicit Alias
descriptor and replaced the previous `named` delegate method on
formatters and visitors with an `alias` method.

All delegate and visitor methods now take Descriptors instead of Labels,
so this is a breaking change.