Finish cleanup
The remaining weird thing is how I made it possible for arrays to
require `length > 0` as a condition of being a required array.

Previously, `required` was simply a responsibility of each type, so
arrays were able to define custom semantics. Now, the responsibility of
`required` is fully extracted into the `Required` descriptor, which only
understands that required means `not null`.

To address this, I added an additional `requiredHint` parameter to the
`validate` function, which types can use to define custom required
semantics. Types can ignore this hint if they want "not null" semantics,
which are applied by the `Required` type. Aliases propagate the hint,
but other containers do not (a required container does not imply
required contents and vice versa).

There is probably a better solution.