Restores List(Record) (#10)
Fixes #9

Previously, Records were a special alternative type, and in particular
RecordBuilder didn't implement TypeBuilder.

This commit creates a new, more abstract interface called
TypeBuilderMember, which simply requires that an object that wants to be
allowed to go inside of List() or Dictionary() must provide a builder
property that returns a real builder.

In practice, this means that when you put a Record into a List or
Dictionary, it effectively turns into its equivalent Dictionary for
validation purposes. It is still serialized as an alias, though.