Add support for self-referential types
The short explanation is that as we dehydrate descriptors, we keep a map
of the descriptors we've already seen, and avoid reifying them if they
were already hydrated.

The one wrinkle is that we put a placeholder in the map while we are
hydrating to avoid infinite regress, and that placeholder is not valid
until hydration of the entire root type is complete.

This means that if somehow descriptor hydration interleaves with using
the descriptor with the runtime, you will get a runtime error. But it's
hard to imagine how that would even happen. Let's not do that please?